Beta Restart Update
For this update I had to wipe the slate clean after coming across several hiccups and now have a working stable game again. Will be sure to add more in the future as the weeks close in.

Beta Restart Update

For this update I had to wipe the slate clean after coming across several hiccups and now have a working stable game again. Will be sure to add more in the future as the weeks close in.